python-services/whisper-stt/requirements.txt
# Whisper STT Service Dependencies - Compatible Versions
flask==3.0.3
flask-cors==5.0.0
openai-whisper
numpy<2.0
ffmpeg-python==0.2.0
